refactor : InputView, ResultView 분리
